requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
awscli==1.34.8
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.35.8
botocore==1.35.8
celery==5.4.0
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cytoolz==0.12.3
distro==1.9.0
Django==5.1
django-allauth==64.1.0
django-autoslug==1.9.9
django-bootstrap5==24.2
django-countries==7.6.1
django-crispy-forms==2.3
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-storages==1.13.2
django-utils==0.0.2
djangorestframework==3.15.2
docutils==0.16
eth-hash==0.7.0
eth-typing==5.0.0
eth-utils==5.0.0
frozenlist==1.4.1
getenv==0.2.0
gunicorn==23.0.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.5
httpx==0.27.2
idna==3.7
jiter==0.5.0
jmespath==1.0.1
kombu==5.4.0
multidict==6.0.5
openai==0.28.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pycountry==24.6.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stripe==10.8.0
toolz==0.12.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
